Removed need to sort for wind angle in plot-helper
Previously we sorted the given data for wind angle, in the
plot helper-functions. But since Table (and MultiSails) and Curve
would give sorted data, that would only be necessary for Cloud.

We therefore fixed Clouds get_slices() to also return sorted data
(Might change, so that Cloud is always sorted anyway)